server_py/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
loguru>=0.7.0
google-genai>=0.2.0
reportlab>=4.0.0
email-validator>=2.0.0
alibabacloud-dm20151123>=1.0.0
alembic>=1.13.0
redis>=5.0.0
minio>=7.2.0
oss2>=2.18.0

